requirements.txt
streamlit
pandas
numpy
plotly
pybase64
pymongo
pytz
bcrypt
pymongo[srv]
pydot
graphviz
connect
bs4
nltk
matplotlib
xlrd
openpyxl
webdriver_manager
XlsxWriter
xlrd
pytest
pytest-ordering
lxml
beautifulsoup4
selenium
allure-pytest
scikit-learn